Moved Number monoids to predef
