+++ kategory/src/main/kotlin/kategory/data/Composed.kt
