--- kategory/src/main/kotlin/kategory/instances/ByteMonoid.kt
@@ -1,7 +1,0 @@
-package kategory
-
-object ByteMonoid : Monoid<Byte>, Semigroup<Byte> by SG, GlobalInstance<Monoid<Byte>>() {
-    override fun empty(): Byte = 0
-}
-
-private val SG: Semigroup<Byte> = NumberSemigroup({ one, two -> (one + two).toByte() })
