--- kategory/src/main/kotlin/kategory/instances/DoubleMonoid.kt
@@ -1,7 +1,0 @@
-package kategory
-
-object DoubleMonoid : Monoid<Double>, Semigroup<Double> by SG, GlobalInstance<Monoid<Double>>() {
-    override fun empty(): Double = .0
-}
-
-private val SG: Semigroup<Double> = NumberSemigroup(Double::plus)
