--- kategory/src/main/kotlin/kategory/instances/IntMonoid.kt
@@ -1,7 +1,0 @@
-package kategory
-
-object IntMonoid : Monoid<Int>, Semigroup<Int> by SG, GlobalInstance<Monoid<Int>>() {
-    override fun empty(): Int = 0
-}
-
-private val SG: Semigroup<Int> = NumberSemigroup(Int::plus)
