--- kategory/src/main/kotlin/kategory/instances/ShortMonoid.kt
@@ -1,7 +1,0 @@
-package kategory
-
-object ShortMonoid : Monoid<Short>, Semigroup<Short> by SG, GlobalInstance<Monoid<Short>>() {
-    override fun empty(): Short = 0
-}
-
-private val SG: Semigroup<Short> = NumberSemigroup({ one, two -> (one + two).toShort() })
